move get_completion_signatures into __completion_signatures.hpp
